linux-gen, include: switch packet headers to api+abi
