+++ include/odp/arch/x86_64-linux/odp/api/abi/packet_flags.h
@@ -1,0 +1,7 @@
+/* Copyright (c) 2017, Linaro Limited
+ * All rights reserved.
+ *
+ * SPDX-License-Identifier:     BSD-3-Clause
+ */
+
+#include <odp/api/abi-default/packet_flags.h>
