+++ platform/linux-generic/pktio/dpdk.c
@@ -17,8 +17,8 @@
 
 #include <odp/api/cpumask.h>
 
-#include <odp/api/plat/packet_inlines.h>
 #include <odp/api/packet.h>
+#include <odp/api/plat/packet_inlines.h>
 
 #include <odp_packet_io_internal.h>
 #include <odp_classification_internal.h>
