+++ platform/linux-generic/pktio/netmap.c
@@ -10,8 +10,8 @@
 
 #include <odp_posix_extensions.h>
 
-#include <odp/api/plat/packet_inlines.h>
 #include <odp/api/packet.h>
+#include <odp/api/plat/packet_inlines.h>
 
 #include <odp_packet_io_internal.h>
 #include <odp_packet_netmap.h>
